feat(but): replace old move command with move2
