+++ crates/but/src/command/branch/mod.rs
@@ -1,6 +1,4 @@
 mod apply;
-mod move_branch;
 mod update;
 pub use apply::apply;
-pub(crate) use move_branch::{move_branch_by_name_with_perm, tear_off_branch_by_name_with_perm};
 pub use update::update;
