+++ crates/but/src/command/commit/mod.rs
@@ -1,2 +1,1 @@
 pub mod file;
-pub mod r#move;
