--- crates/but/src/command/commit/move.rs
@@ -1,674 +1,0 @@
-use crate::{
-    CliId,
-    id::{CommitId, CommittedFileId},
-    theme::{self, Paint},
-    utils::OutputChannel,
-};
-use anyhow::bail;
-use but_core::{DryRun, sync::RepoExclusive};
-use but_ctx::Context;
-use but_rebase::graph_rebase::mutate::{InsertSide, RelativeTo};
-
-pub(crate) fn handle_resolved_with_perm(
-    ctx: &mut Context,
-    out: &mut OutputChannel,
-    source_id: &CliId,
-    target_id: &CliId,
-    after: bool,
-    perm: &mut RepoExclusive,
-) -> anyhow::Result<()> {
-    let t = theme::get();
-    // Validate --after flag usage
-    if after {
-        // Check if target is a branch (--after only makes sense for commit-to-commit moves)
-        if matches!(target_id, CliId::Branch(..)) {
-            bail!(
-                "The {} flag only makes sense when moving a commit to another commit.\n\
-                When moving to a branch, the commit is placed at the top of the stack by default.",
-                "--after"
-            );
-        }
-        // Check if source is a committed file (--after doesn't make sense for file moves)
-        if matches!(source_id, CliId::CommittedFile { .. }) {
-            bail!(
-                "The {} flag only makes sense when moving a commit to another commit.\n\
-                When moving a file from one commit to another, the changes are simply transferred.",
-                "--after"
-            );
-        }
-    }
-
-    // Route and execute the operation
-    let Some(operation) = route_move_operation(source_id, target_id, after) else {
-        bail!(
-            "Cannot move {} ({}) to {} ({}).\n\
-            Valid moves: commit→commit, commit→branch, or committed-file→commit",
-            t.cli_id.paint(source_id.to_short_string()),
-            t.attention.paint(source_id.kind_for_humans()),
-            t.cli_id.paint(target_id.to_short_string()),
-            t.attention.paint(target_id.kind_for_humans())
-        );
-    };
-
-    operation.execute_with_perm(ctx, out, perm)
-}
-
-pub(crate) fn handle_multiple_resolved_with_perm(
-    ctx: &mut Context,
-    out: &mut OutputChannel,
-    source_ids: &[CliId],
-    target_id: &CliId,
-    after: bool,
-    perm: &mut RepoExclusive,
-) -> anyhow::Result<()> {
-    let t = theme::get();
-    if source_ids.is_empty() {
-        bail!("No source commits provided.");
-    }
-
-    let sources = source_ids
-        .iter()
-        .map(|id| match id {
-            CliId::Commit(CommitId { commit_id, .. }) => Ok(*commit_id),
-            _ => bail!(
-                "Cannot move {} as part of a multi-commit operation. Only commits are supported.",
-                t.cli_id.paint(id.to_short_string())
-            ),
-        })
-        .collect::<anyhow::Result<Vec<_>>>()?;
-
-    match target_id {
-        CliId::Commit(CommitId { commit_id, .. }) => {
-            move_commit_to_commit_with_perm(ctx, sources, *commit_id, after, out, perm)
-        }
-        CliId::Branch(branch) => {
-            if after {
-                bail!(
-                    "The {} flag only makes sense when moving a commit to another commit.\n\
-                    When moving to a branch, the commit is placed at the top of the stack by default.",
-                    "--after"
-                );
-            }
-            move_commit_to_branch_with_perm(ctx, sources, &branch.name, out, perm)
-        }
-        _ => bail!(
-            "Cannot move multiple commits to {} ({}).\n\
-            Valid multi-commit moves: commit→commit or commit→branch",
-            t.cli_id.paint(target_id.to_short_string()),
-            t.attention.paint(target_id.kind_for_humans())
-        ),
-    }
-}
-
-/// Represents the operation to perform for a given source and target combination in `but move`.
-#[derive(Debug)]
-enum MoveOperation<'a> {
-    /// Move a commit to be before/after another commit
-    CommitToCommit {
-        source: gix::ObjectId,
-        target: gix::ObjectId,
-        after: bool,
-    },
-    /// Move a commit to a branch (places at top of the branch)
-    CommitToBranch {
-        source: gix::ObjectId,
-        target_branch: &'a str,
-    },
-    /// Move a committed file to another commit (delegates to rub)
-    CommittedFileToCommit {
-        path: &'a bstr::BStr,
-        source_commit: gix::ObjectId,
-        target_commit: gix::ObjectId,
-    },
-}
-
-impl<'a> MoveOperation<'a> {
-    /// Executes this move operation with an existing exclusive permission token.
-    fn execute_with_perm(
-        self,
-        ctx: &mut Context,
-        out: &mut OutputChannel,
-        perm: &mut RepoExclusive,
-    ) -> anyhow::Result<()> {
-        match self {
-            MoveOperation::CommitToCommit {
-                source,
-                target,
-                after,
-            } => move_commit_to_commit_with_perm(ctx, vec![source], target, after, out, perm),
-            MoveOperation::CommitToBranch {
-                source,
-                target_branch,
-            } => move_commit_to_branch_with_perm(ctx, vec![source], target_branch, out, perm),
-            MoveOperation::CommittedFileToCommit {
-                path,
-                source_commit,
-                target_commit,
-            } => super::file::commited_file_to_another_commit_with_perm(
-                ctx,
-                path,
-                source_commit,
-                target_commit,
-                out,
-                perm,
-            ),
-        }
-    }
-}
-
-/// Move a commit to a new position relative to another one.
-pub fn move_commit_to_commit_with_perm(
-    ctx: &mut Context,
-    sources: Vec<gix::ObjectId>,
-    target: gix::ObjectId,
-    after: bool,
-    out: &mut OutputChannel,
-    perm: &mut RepoExclusive,
-) -> Result<(), anyhow::Error> {
-    let side = if after {
-        InsertSide::Above
-    } else {
-        InsertSide::Below
-    };
-
-    if sources.contains(&target) {
-        if let Some(out) = out.for_human() {
-            if sources.len() == 1 {
-                writeln!(out, "Source and target are the same commit. Nothing to do.")?;
-            } else {
-                writeln!(
-                    out,
-                    "At least one source commit is the same as the target commit. Nothing to do."
-                )?;
-            }
-        } else if let Some(out) = out.for_json() {
-            out.write_value(serde_json::json!({"ok": true}))?;
-        }
-        return Ok(());
-    }
-
-    let (source_change_ids, target_change_id) = {
-        let repo = ctx.repo.get()?;
-        let source_change_ids = sources
-            .iter()
-            .map(|id| crate::utils::get_change_id_for_commit(&repo, *id))
-            .collect::<anyhow::Result<Vec<_>>>()?;
-        let target_change_id = crate::utils::get_change_id_for_commit(&repo, target)?;
-        (source_change_ids, target_change_id)
-    };
-
-    let result = but_api::commit::move_commit::commit_move_with_perm(
-        ctx,
-        sources.clone(),
-        RelativeTo::Commit(target),
-        side,
-        DryRun::No,
-        perm,
-    )?;
-
-    let t = theme::get();
-    if let Some(out) = out.for_human() {
-        let action = if after { "after" } else { "before" };
-        if sources.len() == 1 {
-            let source = theme::Commit(
-                post_move_id(&result.workspace, sources[0]),
-                Some(source_change_ids[0].clone()),
-            );
-            let target = theme::Commit(
-                post_move_id(&result.workspace, target),
-                Some(target_change_id),
-            );
-            writeln!(out, "Moved {source} → {action} {target}")?;
-        } else {
-            let target = theme::Commit(
-                post_move_id(&result.workspace, target),
-                Some(target_change_id),
-            );
-            writeln!(
-                out,
-                "Moved {} commits → {action} {target}",
-                t.cli_id.paint(sources.len().to_string()),
-            )?;
-        }
-    } else if let Some(out) = out.for_json() {
-        out.write_value(serde_json::json!({"ok": true}))?;
-    }
-    Ok(())
-}
-
-/// The id a commit has after the move; commits untouched by the rebase keep
-/// their id.
-fn post_move_id(workspace: &but_api::WorkspaceState, id: gix::ObjectId) -> gix::ObjectId {
-    workspace.replaced_commits.get(&id).copied().unwrap_or(id)
-}
-
-pub fn move_commit_to_branch_with_perm(
-    ctx: &mut Context,
-    sources: Vec<gix::ObjectId>,
-    target_branch: &str,
-    out: &mut OutputChannel,
-    perm: &mut RepoExclusive,
-) -> Result<(), anyhow::Error> {
-    let target_full_name = gix::refs::FullName::try_from(format!("refs/heads/{target_branch}"))?;
-    let source_change_id = if sources.len() == 1 {
-        let repo = ctx.repo.get()?;
-        Some(crate::utils::get_change_id_for_commit(&repo, sources[0])?)
-    } else {
-        None
-    };
-    let result = but_api::commit::move_commit::commit_move_with_perm(
-        ctx,
-        sources.clone(),
-        RelativeTo::Reference(target_full_name),
-        InsertSide::Below,
-        DryRun::No,
-        perm,
-    )?;
-
-    let t = theme::get();
-    if let Some(out) = out.for_human() {
-        if sources.len() == 1 {
-            let source = theme::Commit(
-                post_move_id(&result.workspace, sources[0]),
-                source_change_id,
-            );
-            writeln!(
-                out,
-                "Moved {} → {}",
-                source,
-                t.local_branch.paint(format!("[{target_branch}]"))
-            )?;
-        } else {
-            writeln!(
-                out,
-                "Moved {} commits → {}",
-                t.cli_id.paint(sources.len().to_string()),
-                t.local_branch.paint(format!("[{target_branch}]"))
-            )?;
-        }
-    } else if let Some(out) = out.for_json() {
-        out.write_value(serde_json::json!({"ok": true}))?;
-    }
-    Ok(())
-}
-
-/// Determines the move operation to perform for a given source and target combination.
-/// Returns `Some(operation)` if the combination is valid, `None` otherwise.
-fn route_move_operation<'a>(
-    source: &'a CliId,
-    target: &'a CliId,
-    after: bool,
-) -> Option<MoveOperation<'a>> {
-    use CliId::*;
-
-    match (source, target) {
-        // Commit -> Commit: move commit to specific position
-        (
-            Commit(CommitId {
-                commit_id: source, ..
-            }),
-            Commit(CommitId {
-                commit_id: target, ..
-            }),
-        ) => Some(MoveOperation::CommitToCommit {
-            source: *source,
-            target: *target,
-            after,
-        }),
-        // Commit -> Branch: move commit to top of branch
-        (
-            Commit(CommitId {
-                commit_id: source, ..
-            }),
-            Branch(branch),
-        ) => Some(MoveOperation::CommitToBranch {
-            source: *source,
-            target_branch: &branch.name,
-        }),
-        // CommittedFile -> Commit: move a file from one commit to another
-        (
-            CommittedFile(CommittedFileId {
-                path,
-                commit_id: source_commit,
-                ..
-            }),
-            Commit(CommitId {
-                commit_id: target_commit,
-                ..
-            }),
-        ) => Some(MoveOperation::CommittedFileToCommit {
-            path: path.as_ref(),
-            source_commit: *source_commit,
-            target_commit: *target_commit,
-        }),
-        // All other combinations are invalid for move
-        _ => None,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use bstr::BString;
-
-    use crate::id::{BranchId, CommitId, WorktreeHunk};
-
-    use super::*;
-
-    // Helper to create test CliIds
-    fn commit_id(id: &str) -> CliId {
-        CliId::Commit(CommitId {
-            commit_id: gix::ObjectId::empty_tree(gix::hash::Kind::Sha1),
-            id: id.to_string(),
-            change_id: None,
-        })
-    }
-
-    fn branch_id(name: &str) -> CliId {
-        CliId::Branch(BranchId {
-            name: name.to_string(),
-            id: "br".to_string(),
-            stack_id: None,
-        })
-    }
-
-    fn uncommitted_id() -> CliId {
-        CliId::UncommittedHunkOrFile(crate::id::UncommittedHunkOrFile {
-            id: "uc".to_string(),
-            hunk_assignments: nonempty::NonEmpty::new(WorktreeHunk {
-                id: None,
-                hunk_header: None,
-                path: "test.txt".to_string(),
-                path_bytes: BString::from("test.txt"),
-                line_nums_added: None,
-                line_nums_removed: None,
-                diff: None,
-            }),
-            is_entire_file: true,
-        })
-    }
-
-    fn committed_file_id(path: &str) -> CliId {
-        CliId::CommittedFile(CommittedFileId {
-            commit_id: gix::ObjectId::empty_tree(gix::hash::Kind::Sha1),
-            path: BString::from(path),
-            id: "cf".to_string(),
-            change_id: None,
-        })
-    }
-
-    #[test]
-    fn test_route_move_operation_valid_combinations() {
-        // Commit -> Commit: should route to CommitToCommit
-        let source = commit_id("a");
-        let target = commit_id("b");
-        let result = route_move_operation(&source, &target, false);
-        assert!(matches!(result, Some(MoveOperation::CommitToCommit { .. })));
-
-        // Commit -> Commit with --after: should route to CommitToCommit with after=true
-        let source = commit_id("a");
-        let target = commit_id("b");
-        let result = route_move_operation(&source, &target, true);
-        assert!(matches!(
-            result,
-            Some(MoveOperation::CommitToCommit { after: true, .. })
-        ));
-
-        // Commit -> Branch: should route to CommitToBranch
-        let source = commit_id("a");
-        let target = branch_id("main");
-        let result = route_move_operation(&source, &target, false);
-        assert!(matches!(result, Some(MoveOperation::CommitToBranch { .. })));
-
-        // CommittedFile -> Commit: should route to CommittedFileToCommit
-        let source = committed_file_id("file.txt");
-        let target = commit_id("b");
-        let result = route_move_operation(&source, &target, false);
-        assert!(matches!(
-            result,
-            Some(MoveOperation::CommittedFileToCommit { .. })
-        ));
-    }
-
-    #[test]
-    fn test_route_move_operation_invalid_combinations() {
-        // Uncommitted -> Commit: not supported by move
-        let source = uncommitted_id();
-        let target = commit_id("a");
-        let result = route_move_operation(&source, &target, false);
-        assert!(result.is_none());
-
-        // Commit -> Uncommitted: not supported by move
-        let source = commit_id("a");
-        let target = uncommitted_id();
-        let result = route_move_operation(&source, &target, false);
-        assert!(result.is_none());
-
-        // Branch -> Commit: not supported by move (this is a rub operation)
-        let source = branch_id("main");
-        let target = commit_id("a");
-        let result = route_move_operation(&source, &target, false);
-        assert!(result.is_none());
-
-        // CommittedFile -> Branch: not supported by move (use rub for this)
-        let source = committed_file_id("file.txt");
-        let target = branch_id("main");
-        let result = route_move_operation(&source, &target, false);
-        assert!(result.is_none());
-
-        // CommittedFile -> Uncommitted: not supported by move
-        let source = committed_file_id("file.txt");
-        let target = uncommitted_id();
-        let result = route_move_operation(&source, &target, false);
-        assert!(result.is_none());
-    }
-
-    /// Test the index adjustment logic when moving commits within the same series.
-    /// This is the most complex part of the move logic and most prone to off-by-one errors.
-    #[test]
-    fn test_index_adjustment_logic() {
-        // Simulating the logic from move_within_stack without needing a full Git repo
-
-        // Test case 1: Moving from earlier to later position (source_pos < target_pos)
-        // Initial: [A, B, C, D] at indices [0, 1, 2, 3]
-        // Move A (index 0) before D (index 3)
-        let source_pos = 0;
-        let target_pos = 3;
-        let source_idx = 0;
-        let target_idx = 0; // same series
-
-        // After removal: [B, C, D] at indices [0, 1, 2]
-        // Target D is now at index 2 (was 3, minus 1 for removal)
-        let adjusted_target_pos = if source_idx == target_idx && source_pos < target_pos {
-            target_pos - 1
-        } else {
-            target_pos
-        };
-        assert_eq!(adjusted_target_pos, 2);
-
-        // Insert before D means insert at adjusted_target_pos + 1
-        let insert_position = adjusted_target_pos + 1; // after=false case
-        assert_eq!(insert_position, 3);
-
-        // Result: [B, C, D, A] - A is now before (older than) D ✓
-
-        // Test case 2: Moving from later to earlier position (source_pos > target_pos)
-        // Initial: [A, B, C, D] at indices [0, 1, 2, 3]
-        // Move D (index 3) after A (index 0)
-        let source_pos = 3;
-        let target_pos = 0;
-        let source_idx = 0;
-        let target_idx = 0; // same series
-
-        // After removal: [A, B, C] at indices [0, 1, 2]
-        // Target A stays at index 0 (no adjustment needed)
-        let adjusted_target_pos = if source_idx == target_idx && source_pos < target_pos {
-            target_pos - 1
-        } else {
-            target_pos
-        };
-        assert_eq!(adjusted_target_pos, 0);
-
-        // Insert after A means insert at adjusted_target_pos
-        let insert_position = adjusted_target_pos; // after=true case
-        assert_eq!(insert_position, 0);
-
-        // Result: [D, A, B, C] - D is now after (newer than) A ✓
-
-        // Test case 3: Moving within adjacent positions
-        // Initial: [A, B, C] at indices [0, 1, 2]
-        // Move B (index 1) before C (index 2)
-        let source_pos = 1;
-        let target_pos = 2;
-        let source_idx = 0;
-        let target_idx = 0;
-
-        // After removal: [A, C] at indices [0, 1]
-        let adjusted_target_pos = if source_idx == target_idx && source_pos < target_pos {
-            target_pos - 1
-        } else {
-            target_pos
-        };
-        assert_eq!(adjusted_target_pos, 1);
-
-        // Insert before C means insert at adjusted_target_pos + 1
-        let insert_position = adjusted_target_pos + 1;
-        assert_eq!(insert_position, 2);
-
-        // Result: [A, C, B] - B is now before (older than) C ✓
-    }
-
-    #[test]
-    fn test_index_adjustment_different_series() {
-        // When moving between different series, no adjustment needed
-        // Initial series 0: [A, B, C]
-        // Initial series 1: [D, E, F]
-        // Move B (series 0, index 1) to before E (series 1, index 1)
-
-        let source_pos = 1;
-        let target_pos = 1;
-        let source_idx = 0;
-        let target_idx = 1; // different series
-
-        // After removal from series 0: [A, C]
-        // Series 1 is unchanged: [D, E, F]
-        // No adjustment needed because they're in different series
-        let adjusted_target_pos = if source_idx == target_idx && source_pos < target_pos {
-            target_pos - 1
-        } else {
-            target_pos
-        };
-        assert_eq!(adjusted_target_pos, 1); // No adjustment
-
-        // Insert before E means insert at adjusted_target_pos + 1
-        let insert_position = adjusted_target_pos + 1;
-        assert_eq!(insert_position, 2);
-
-        // Result series 1: [D, E, B, F] - B is now before (older than) E ✓
-    }
-
-    #[test]
-    fn test_before_vs_after_semantics() {
-        // Verify the "before" (default) and "after" semantics are correct
-        // Given commits ordered newest to oldest: [NEW, MID, OLD]
-        // at indices [0, 1, 2]
-
-        // "Move X after Y" means X becomes newer than Y (X depends on Y)
-        // So moving MID after NEW should give [MID, NEW, OLD]
-        let source_pos = 1; // MID
-        let target_pos = 0; // NEW
-        let after = true;
-
-        let adjusted_target_pos = if source_pos < target_pos {
-            target_pos - 1
-        } else {
-            target_pos
-        };
-        let insert_position = if after {
-            adjusted_target_pos
-        } else {
-            adjusted_target_pos + 1
-        };
-        assert_eq!(insert_position, 0); // MID goes to index 0, pushing NEW down
-
-        // "Move X before Y" means X becomes older than Y (Y depends on X)
-        // So moving MID before OLD should give [NEW, OLD, MID]
-        let source_pos = 1; // MID
-        let target_pos = 2; // OLD
-        let after = false;
-
-        let adjusted_target_pos = if source_pos < target_pos {
-            target_pos - 1
-        } else {
-            target_pos
-        };
-        let insert_position = if after {
-            adjusted_target_pos
-        } else {
-            adjusted_target_pos + 1
-        };
-        assert_eq!(adjusted_target_pos, 1); // OLD moved to index 1 after removing MID
-        assert_eq!(insert_position, 2); // MID goes after OLD (higher index = older)
-    }
-
-    #[test]
-    fn test_moving_to_same_position_logic() {
-        // When moving a commit and it ends up at the same logical position,
-        // we want to ensure the math works out correctly
-
-        // Move C after B when they're already in order [A, B, C]
-        // This should be essentially a no-op after the adjustment
-        let source_pos = 2; // C
-        let target_pos = 1; // B
-        let after = true; // after B
-
-        // After removing C: [A, B] at [0, 1]
-        // B is still at index 1
-        let adjusted_target_pos = if source_pos < target_pos {
-            target_pos - 1
-        } else {
-            target_pos
-        };
-        assert_eq!(adjusted_target_pos, 1);
-
-        // Insert after B (index 1) means insert at index 1
-        let insert_position = if after {
-            adjusted_target_pos
-        } else {
-            adjusted_target_pos + 1
-        };
-        assert_eq!(insert_position, 1);
-
-        // Result: [A, C, B] - Wait, this is wrong for a no-op case!
-        // Actually this isn't a no-op - "after B" means newer than B
-        // So C moves to be newer than B, which is a real change.
-    }
-
-    #[test]
-    fn test_edge_case_two_commits() {
-        // Test with only two commits to ensure boundary conditions work
-
-        // [A, B] - Move B before A (swap them)
-        let source_pos = 1;
-        let target_pos = 0;
-        let after = false;
-
-        let adjusted_target_pos = if source_pos < target_pos {
-            target_pos - 1
-        } else {
-            target_pos
-        };
-        assert_eq!(adjusted_target_pos, 0);
-
-        let insert_position = if after {
-            adjusted_target_pos
-        } else {
-            adjusted_target_pos + 1
-        };
-        assert_eq!(insert_position, 1);
-
-        // Result: [A, B] - B stays in same position, which means B is still newer
-        // Actually "before A" means older than A, so B should go after A
-        // But [A, B] with B at index 1 means B is older than A already
-        // So this is actually correct - B is before (older than) A
-    }
-}
