+++ crates/but/src/command/mod.rs
@@ -13,7 +13,6 @@ pub(crate) mod external;
 pub(crate) mod git_config;
 pub mod gui;
 pub mod help;
-pub mod r#move;
 pub mod onboarding;
 pub mod open;
 pub mod push;
