--- crates/but/src/command/move.rs
@@ -1,122 +1,0 @@
-use anyhow::Context as _;
-
-use crate::{CliId, IdMap, utils::OutputChannel};
-
-pub(crate) fn handle(
-    ctx: &mut but_ctx::Context,
-    out: &mut OutputChannel,
-    source: &str,
-    target: &str,
-    after: bool,
-) -> anyhow::Result<()> {
-    let mut guard = ctx.exclusive_worktree_access();
-    let id_map = IdMap::new_from_context(ctx, None, guard.read_permission())?;
-    let source_ids =
-        resolve_many(&id_map, ctx, source, "Source").context("Failed to move commit.")?;
-    let target_id =
-        resolve_single(&id_map, ctx, target, "Target").context("Failed to move commit.")?;
-
-    if source_ids.is_empty() {
-        anyhow::bail!("Source must include at least one selector")
-    }
-
-    if source_ids.len() > 1 {
-        return super::commit::r#move::handle_multiple_resolved_with_perm(
-            ctx,
-            out,
-            &source_ids,
-            &target_id,
-            after,
-            guard.write_permission(),
-        )
-        .context("Failed to move commit.");
-    }
-
-    let source_id = source_ids[0].clone();
-
-    let branch_route = matches!(
-        (&source_id, &target_id),
-        (
-            CliId::Branch(..),
-            CliId::Branch(..) | CliId::Uncommitted { .. }
-        )
-    );
-
-    let move_result = if branch_route {
-        let (CliId::Branch(source), target_id) = (&source_id, &target_id) else {
-            unreachable!("branch_route guarantees source is a branch")
-        };
-        if after {
-            // TODO: Allow to move branch after another below another branch.
-            Err(anyhow::anyhow!(
-                "The --after flag only makes sense when moving a commit to another commit."
-            ))
-        } else {
-            match target_id {
-                CliId::Uncommitted { .. } => super::branch::tear_off_branch_by_name_with_perm(
-                    ctx,
-                    &source.name,
-                    out,
-                    guard.write_permission(),
-                ),
-                CliId::Branch(target) => super::branch::move_branch_by_name_with_perm(
-                    ctx,
-                    &source.name,
-                    &target.name,
-                    out,
-                    guard.write_permission(),
-                ),
-                _ => unreachable!("branch_route guarantees target is branch or uncommitted"),
-            }
-        }
-    } else {
-        super::commit::r#move::handle_resolved_with_perm(
-            ctx,
-            out,
-            &source_id,
-            &target_id,
-            after,
-            guard.write_permission(),
-        )
-    };
-
-    if branch_route {
-        move_result.context("Failed to move branch.")
-    } else {
-        move_result.context("Failed to move commit.")
-    }
-}
-
-fn resolve_single(
-    id_map: &IdMap,
-    ctx: &but_ctx::Context,
-    selector: &str,
-    kind: &str,
-) -> anyhow::Result<CliId> {
-    let matches = id_map.parse_using_context(selector, ctx)?;
-    if matches.is_empty() {
-        anyhow::bail!(
-            "{kind} '{selector}' not found. If you just performed a Git operation, try running 'but status' to refresh."
-        );
-    }
-    if matches.len() > 1 {
-        anyhow::bail!(
-            "{kind} '{selector}' is ambiguous. Try using more characters to disambiguate."
-        );
-    }
-    Ok(matches[0].clone())
-}
-
-fn resolve_many(
-    id_map: &IdMap,
-    ctx: &but_ctx::Context,
-    selectors: &str,
-    kind: &str,
-) -> anyhow::Result<Vec<CliId>> {
-    selectors
-        .split(',')
-        .map(str::trim)
-        .filter(|selector| !selector.is_empty())
-        .map(|selector| resolve_single(id_map, ctx, selector, kind))
-        .collect()
-}
