+++ crates/but/tests/but/command/move.rs
@@ -1,169 +1,132 @@
-use anyhow::Context as _;
-use but_core::{
-    RefMetadata,
-    ref_metadata::{StackId, WorkspaceCommitRelation, WorkspaceStack, WorkspaceStackBranch},
-};
-use snapbox::str;
-
 use crate::{
-    command::util::{commit_two_files_as_two_hunks_each, find_branch},
-    utils::{CommandExt, Sandbox},
+    command::util::{
+        branch_commit_cli_ids, commit_two_files_as_two_hunks_each,
+        status_json_with_files as status_json,
+    },
+    utils::Sandbox,
 };
 
 #[test]
-fn move_commit_before_another_commit() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-
+fn move_commit_above_other_commit() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
     env.setup_metadata(&["A"]);
 
-    // Create three commits
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "c.txt", "d.txt", "second commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "e.txt", "f.txt", "third commit");
-
-    // Get commit CLI IDs from status
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commits = &status_json["stacks"][0]["branches"][0]["commits"];
-
-    // Commits are ordered newest first, so:
-    // commits[0] = "third commit"
-    // commits[1] = "second commit"
-    // commits[2] = "first commit"
-    // commits[3] = initial "add A" commit
-
-    let third_commit_cli_id = commits[0]["cliId"].as_str().unwrap();
-    let first_commit_cli_id = commits[2]["cliId"].as_str().unwrap();
-
-    // Move "third commit" before "first commit" (making it the oldest)
-    env.but(format!("move {third_commit_cli_id} {first_commit_cli_id}"))
+    env.but("status")
         .assert()
         .success()
-        .stdout_eq(str![[r#"
-Moved 1 → before 1
-
-"#]]);
-
-    // Verify the move was successful by checking that we still have the right number of commits
-    // (The actual reordering is tested in unit tests, here we just verify the command executed)
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commits_after = &status_json["stacks"][0]["branches"][0]["commits"];
-
-    // Should still have 4 commits (3 we created + initial "add A")
-    assert_eq!(commits_after.as_array().unwrap().len(), 4);
-
-    Ok(())
-}
-
-#[test]
-fn move_commit_after_another_commit() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-
-    env.setup_metadata(&["A"]);
-
-    // Create three commits
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "c.txt", "d.txt", "second commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "e.txt", "f.txt", "third commit");
-
-    // Get commit CLI IDs
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commits = &status_json["stacks"][0]["branches"][0]["commits"];
-
-    let first_commit_cli_id = commits[2]["cliId"].as_str().unwrap();
-    let third_commit_cli_id = commits[0]["cliId"].as_str().unwrap();
-
-    // Move "first commit" after "third commit" (making it the newest)
-    env.but(format!(
-        "move {first_commit_cli_id} {third_commit_cli_id} --after"
-    ))
-    .assert()
-    .success()
-    .stdout_eq(str![[r#"
-Moved 1 → after 1
-
-"#]]);
-
-    // Verify the move was successful by checking that we still have the right number of commits
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commits_after = &status_json["stacks"][0]["branches"][0]["commits"];
-
-    // Should still have 4 commits (3 we created + initial "add A")
-    assert_eq!(commits_after.as_array().unwrap().len(), 4);
-
-    Ok(())
-}
-
-#[test]
-fn move_multiple_commits_before_another_commit() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-
-    env.setup_metadata(&["A"]);
-
-    // Create three commits
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "c.txt", "d.txt", "second commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "e.txt", "f.txt", "third commit");
-
-    env.but("st")
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
         .stdout_eq(snapbox::str![[r#"
 ╭┄ zz [uncommitted] (no changes)
 ┊
 ┊╭┄ g0 [A]
-┊●   1#0 create e.txt and f.txt
-┊●   1#1 create c.txt and d.txt
-┊●   1#2 create a.txt and b.txt
-┊●   tpm add A
+┊●   ywx add second
+┊●   zll add first
 ├╯
 ┊
-┴ 0dc3733 (common base) 2000-01-02 add M
+┴ 1bbc04b (common base) 2000-01-02 add Base
 
 Hint: run `but help` for all commands
 
 "#]]);
-    // Commits are ordered newest first.
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commits = &status_json["stacks"][0]["branches"][0]["commits"];
 
-    let third_commit_cli_id = commits[0]["cliId"].as_str().unwrap();
-    let second_commit_cli_id = commits[1]["cliId"].as_str().unwrap();
-    let first_commit_cli_id = commits[2]["cliId"].as_str().unwrap();
-
-    env.but(format!(
-        "move {second_commit_cli_id},{third_commit_cli_id} {first_commit_cli_id}"
-    ))
-    .assert()
-    .success()
-    .stdout_eq(str![[r#"
-Moved 2 commits → before 1
+    env.but("move zll --above ywx")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved fe12bcd above commit 9ac4652
 
 "#]]);
 
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commits_after = &status_json["stacks"][0]["branches"][0]["commits"];
-
-    assert_eq!(commits_after.as_array().unwrap().len(), 4);
-
-    env.but("st")
+    env.but("status")
         .assert()
         .success()
-        .stderr_eq(snapbox::str![])
         .stdout_eq(snapbox::str![[r#"
 ╭┄ zz [uncommitted] (no changes)
 ┊
 ┊╭┄ g0 [A]
-┊●   1#0 create a.txt and b.txt
-┊●   1#1 create e.txt and f.txt
-┊●   1#2 create c.txt and d.txt
-┊●   tpm add A
+┊●   zll add first
+┊●   ywx add second
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn move_commit_below_other_commit() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
+    env.setup_metadata(&["A"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊●   zll add first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move ywx --below zll")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 9ac4652 below commit fe12bcd
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   zll add first
+┊●   ywx add second
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn move_multiple_consecutive_commits_relative_to_other_commit() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("commits-with-same-prefix");
+    env.setup_metadata(&["A"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   usn add A13
+┊●   opy add A12
+┊●   opk add A11
+┊●   vvl add A10
+┊●   mzz add A9
+┊●   vmw add A8
+┊●   tpw add A7
+┊●   lyq add A6
+┊●   pyq add A5
+┊●   mvv add A4
+┊●   tvm add A3
+┊●   sxq add A2
+┊●   zpl add A1
 ├╯
 ┊
 ┴ 0dc3733 (common base) 2000-01-02 add M
@@ -172,76 +135,37 @@ Hint: run `but help` for all commands
 
 "#]]);
 
-    Ok(())
-}
+    for (operator, target_cli_id) in [("--above", "lyq"), ("--below", "tpw")] {
+        env.but("move vvl mzz")
+            .arg(operator)
+            .arg(target_cli_id)
+            .assert()
+            .success()
+            .stdout_eq(snapbox::str![["
+Moved c472887, 8188106 [..]
 
-#[test]
-fn move_multiple_commits_after_another_commit() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
+"]]);
 
-    env.setup_metadata(&["A"]);
-
-    // Create three commits
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "c.txt", "d.txt", "second commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "e.txt", "f.txt", "third commit");
-
-    env.but("st")
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
+        env.but("status")
+            .assert()
+            .success()
+            .stdout_eq(snapbox::str![[r#"
 ╭┄ zz [uncommitted] (no changes)
 ┊
 ┊╭┄ g0 [A]
-┊●   1#0 create e.txt and f.txt
-┊●   1#1 create c.txt and d.txt
-┊●   1#2 create a.txt and b.txt
-┊●   tpm add A
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
-
-"#]]);
-    // Commits are ordered newest first.
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commits = &status_json["stacks"][0]["branches"][0]["commits"];
-
-    let third_commit_cli_id = commits[0]["cliId"].as_str().unwrap();
-    let second_commit_cli_id = commits[1]["cliId"].as_str().unwrap();
-    let first_commit_cli_id = commits[2]["cliId"].as_str().unwrap();
-
-    env.but(format!(
-        "move {first_commit_cli_id},{second_commit_cli_id} {third_commit_cli_id} --after"
-    ))
-    .assert()
-    .success()
-    .stdout_eq(str![[r#"
-Moved 2 commits → after 1
-
-"#]]);
-
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commits_after = &status_json["stacks"][0]["branches"][0]["commits"];
-
-    assert_eq!(commits_after.as_array().unwrap().len(), 4);
-
-    env.but("st")
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ g0 [A]
-┊●   1#0 create c.txt and d.txt
-┊●   1#1 create a.txt and b.txt
-┊●   1#2 create e.txt and f.txt
-┊●   tpm add A
+┊●   usn add A13
+┊●   opy add A12
+┊●   opk add A11
+┊●   vmw add A8
+┊●   tpw add A7
+┊●   vvl add A10
+┊●   mzz add A9
+┊●   lyq add A6
+┊●   pyq add A5
+┊●   mvv add A4
+┊●   tvm add A3
+┊●   sxq add A2
+┊●   zpl add A1
 ├╯
 ┊
 ┴ 0dc3733 (common base) 2000-01-02 add M
@@ -250,545 +174,399 @@ Hint: run `but help` for all commands
 
 "#]]);
 
-    Ok(())
-}
-
-#[test]
-fn move_multiple_commits_from_different_branches() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("three-stacks");
-
-    env.setup_metadata(&["A", "B", "C"]);
-
-    env.but("st")
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ g0 [A]
-┊●   psr A: 10 lines on top
-├╯
-┊
-┊╭┄ h0 [B]
-┊●   ynm B: another 10 lines at the bottom
-┊●   tlv B: 10 lines at the bottom
-├╯
-┊
-┊╭┄ i0 [C]
-┊●   zmt C: add another 10 lines to new file
-┊●   xkt C: add 10 lines to new file
-┊●   sxz C: new file with 10 lines
-├╯
-┊
-┴ 8f0d338 (common base) 2000-01-02 base
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    let status_before = status_json(&env)?;
-
-    let branch_a = find_branch(&status_before, "A")?;
-    let branch_b = find_branch(&status_before, "B")?;
-    let branch_c = find_branch(&status_before, "C")?;
-
-    let a_commits = branch_a["commits"]
-        .as_array()
-        .context("Missing commits for branch A")?;
-    let b_commits = branch_b["commits"]
-        .as_array()
-        .context("Missing commits for branch B")?;
-    let c_commits = branch_c["commits"]
-        .as_array()
-        .context("Missing commits for branch C")?;
-
-    let mut source_commit_cli_ids = Vec::with_capacity(4);
-    source_commit_cli_ids.push(
-        a_commits[0]["cliId"]
-            .as_str()
-            .context("Missing cliId for A commit")?,
-    );
-    for commit in c_commits {
-        source_commit_cli_ids.push(
-            commit["cliId"]
-                .as_str()
-                .context("Missing cliId for C commit")?,
-        );
+        env.but("undo").assert().success();
     }
-
-    let b_tip_commit_cli_id = b_commits[0]["cliId"]
-        .as_str()
-        .context("Missing cliId for B tip commit")?;
-
-    env.but(format!(
-        "move {} {b_tip_commit_cli_id}",
-        source_commit_cli_ids.join(",")
-    ))
-    .assert()
-    .success()
-    .stdout_eq(str![[r#"
-Moved 4 commits → before [..]
-
-"#]]);
-
-    let status_after = status_json(&env)?;
-    let branch_b_after = find_branch(&status_after, "B")?;
-    let b_commits_after = branch_b_after["commits"]
-        .as_array()
-        .context("Missing commits for branch B after move")?;
-
-    assert_eq!(
-        b_commits_after.len(),
-        6,
-        "B should contain its original 2 commits plus the 4 moved commits"
-    );
-
-    let b_messages_after = b_commits_after
-        .iter()
-        .map(|commit| {
-            commit["message"]
-                .as_str()
-                .map(|message| message.trim_end())
-                .context("Missing commit message")
-        })
-        .collect::<anyhow::Result<Vec<_>>>()?;
-
-    assert_eq!(
-        b_messages_after,
-        vec![
-            "B: another 10 lines at the bottom",
-            "C: add another 10 lines to new file",
-            "C: add 10 lines to new file",
-            "C: new file with 10 lines",
-            "A: 10 lines on top",
-            "B: 10 lines at the bottom"
-        ]
-    );
-
-    env.but("st")
-        .assert()
-        .success()
-        .stderr_eq(snapbox::str![])
-        .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ g0 [A] (no commits)
-├╯
-┊
-┊╭┄ h0 [B]
-┊●   [..] B: another 10 lines at the bottom
-┊●   zmt C: add another 10 lines to new file
-┊●   xkt C: add 10 lines to new file
-┊●   sxz C: new file with 10 lines
-┊●   psr A: 10 lines on top
-┊●   tlv B: 10 lines at the bottom
-├╯
-┊
-┊╭┄ i0 [C] (no commits)
-├╯
-┊
-┴ 8f0d338 (common base) 2000-01-02 base
-
-Hint: run `but help` for all commands
-
-"#]]);
-
-    Ok(())
 }
 
 #[test]
-fn move_multiple_commits_from_different_branches_after() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("three-stacks");
+fn move_multiple_non_consecutive_commits_in_arbitrary_order_relative_to_other_commit() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("commits-with-same-prefix");
+    env.setup_metadata(&["A"]);
 
-    env.setup_metadata(&["A", "B", "C"]);
-
-    env.but("st")
+    env.but("status")
         .assert()
         .success()
-        .stderr_eq(snapbox::str![])
         .stdout_eq(snapbox::str![[r#"
 ╭┄ zz [uncommitted] (no changes)
 ┊
 ┊╭┄ g0 [A]
-┊●   psr A: 10 lines on top
+┊●   usn add A13
+┊●   opy add A12
+┊●   opk add A11
+┊●   vvl add A10
+┊●   mzz add A9
+┊●   vmw add A8
+┊●   tpw add A7
+┊●   lyq add A6
+┊●   pyq add A5
+┊●   mvv add A4
+┊●   tvm add A3
+┊●   sxq add A2
+┊●   zpl add A1
 ├╯
 ┊
-┊╭┄ h0 [B]
-┊●   ynm B: another 10 lines at the bottom
-┊●   tlv B: 10 lines at the bottom
-├╯
-┊
-┊╭┄ i0 [C]
-┊●   zmt C: add another 10 lines to new file
-┊●   xkt C: add 10 lines to new file
-┊●   sxz C: new file with 10 lines
-├╯
-┊
-┴ 8f0d338 (common base) 2000-01-02 base
+┴ 0dc3733 (common base) 2000-01-02 add M
 
 Hint: run `but help` for all commands
 
 "#]]);
 
-    let status_before = status_json(&env)?;
+    for (operator, target_cli_id) in [("--above", "vmw"), ("--below", "mzz")] {
+        // We pick the source commits in an "incorrect" order, but they should later be sorted correctly
+        // via topological sort.
+        //
+        // Order as picked is: A7 A1 A5 --above A8, but we expect the commits to be applied from oldest
+        // to newest, i.e. (A8) <- A1 <- A5 <- A7
+        env.but("move tpw zpl pyq")
+            .arg(operator)
+            .arg(target_cli_id)
+            .assert()
+            .success()
+            .stdout_eq(snapbox::str![["
+Moved 2a98cfc, 0748e42, c67c49e [..]
 
-    let branch_a = find_branch(&status_before, "A")?;
-    let branch_b = find_branch(&status_before, "B")?;
-    let branch_c = find_branch(&status_before, "C")?;
+"]]);
 
-    let a_commits = branch_a["commits"]
-        .as_array()
-        .context("Missing commits for branch A")?;
-    let b_commits = branch_b["commits"]
-        .as_array()
-        .context("Missing commits for branch B")?;
-    let c_commits = branch_c["commits"]
-        .as_array()
-        .context("Missing commits for branch C")?;
+        env.but("status")
+            .assert()
+            .success()
+            .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   usn add A13
+┊●   opy add A12
+┊●   opk add A11
+┊●   vvl add A10
+┊●   mzz add A9
+┊●   tpw add A7
+┊●   pyq add A5
+┊●   zpl add A1
+┊●   vmw add A8
+┊●   lyq add A6
+┊●   mvv add A4
+┊●   tvm add A3
+┊●   sxq add A2
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
 
-    let mut source_commit_cli_ids = Vec::with_capacity(4);
-    source_commit_cli_ids.push(
-        a_commits[0]["cliId"]
-            .as_str()
-            .context("Missing cliId for A commit")?,
-    );
-    for commit in c_commits {
-        source_commit_cli_ids.push(
-            commit["cliId"]
-                .as_str()
-                .context("Missing cliId for C commit")?,
-        );
+Hint: run `but help` for all commands
+
+"#]]);
+
+        env.but("undo").assert().success();
     }
+}
 
-    let b_tip_commit_cli_id = b_commits[0]["cliId"]
-        .as_str()
-        .context("Missing cliId for B tip commit")?;
+#[test]
+fn moving_commits_above_branch_creates_branch_above() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
+    env.setup_metadata(&["A"]);
 
-    env.but(format!(
-        "move {} {b_tip_commit_cli_id} --after",
-        source_commit_cli_ids.join(",")
-    ))
-    .assert()
-    .success()
-    .stdout_eq(str![[r#"
-Moved 4 commits → after ynm
-
-"#]]);
-
-    let status_after = status_json(&env)?;
-    let branch_b_after = find_branch(&status_after, "B")?;
-    let b_commits_after = branch_b_after["commits"]
-        .as_array()
-        .context("Missing commits for branch B after move")?;
-
-    assert_eq!(
-        b_commits_after.len(),
-        6,
-        "B should contain its original 2 commits plus the 4 moved commits"
-    );
-
-    let b_messages_after = b_commits_after
-        .iter()
-        .map(|commit| {
-            commit["message"]
-                .as_str()
-                .map(|message| message.trim_end())
-                .context("Missing commit message")
-        })
-        .collect::<anyhow::Result<Vec<_>>>()?;
-
-    assert_eq!(
-        b_messages_after,
-        vec![
-            "C: add another 10 lines to new file",
-            "C: add 10 lines to new file",
-            "C: new file with 10 lines",
-            "A: 10 lines on top",
-            "B: another 10 lines at the bottom",
-            "B: 10 lines at the bottom"
-        ]
-    );
-
-    env.but("st")
+    env.but("status")
         .assert()
         .success()
-        .stderr_eq(snapbox::str![])
         .stdout_eq(snapbox::str![[r#"
 ╭┄ zz [uncommitted] (no changes)
 ┊
-┊╭┄ g0 [A] (no commits)
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊●   zll add first
 ├╯
 ┊
-┊╭┄ h0 [B]
-┊●   zmt C: add another 10 lines to new file
-┊●   xkt C: add 10 lines to new file
-┊●   sxz C: new file with 10 lines
-┊●   psr A: 10 lines on top
-┊●   ynm B: another 10 lines at the bottom
-┊●   tlv B: 10 lines at the bottom
-├╯
-┊
-┊╭┄ i0 [C] (no commits)
-├╯
-┊
-┴ 8f0d338 (common base) 2000-01-02 base
+┴ 1bbc04b (common base) 2000-01-02 add Base
 
 Hint: run `but help` for all commands
 
 "#]]);
 
-    Ok(())
-}
-
-#[test]
-fn move_commit_to_branch() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-
-    // Create a commit on branch A
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "commit on A");
-
-    // Get commit CLI ID
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commit_cli_id = status_json["stacks"][0]["branches"][0]["commits"][0]["cliId"]
-        .as_str()
-        .unwrap();
-
-    // Move commit to branch B
-    env.but(format!("move {commit_cli_id} B"))
+    env.but("move zll --above g0")
         .assert()
         .success()
-        .stdout_eq(str![[r#"
-Moved [..] → [B]
+        .stdout_eq(snapbox::str![[r#"
+Moved fe12bcd to new branch 'a-branch-1' above branch 'A'
 
 "#]]);
 
-    // Verify commit is now on branch B
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ br [a-branch-1]
+┊●   zll add first
+┊│
+┊├┄ g0 [A]
+┊●   ywx add second
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
 
-    // Branch A should have no commits (except the initial one)
-    let branch_a_commits = &status_json["stacks"][0]["branches"][0]["commits"];
-    assert_eq!(branch_a_commits.as_array().unwrap().len(), 1); // Only initial commit
+Hint: run `but help` for all commands
 
-    // Branch B should now have 2 commits (the moved one + initial "add B")
-    let branch_b_commits = &status_json["stacks"][1]["branches"][0]["commits"];
-    assert_eq!(branch_b_commits.as_array().unwrap().len(), 2);
-
-    Ok(())
+"#]]);
 }
 
 #[test]
-fn move_commit_with_invalid_source() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-
+fn moving_commits_above_branch_without_changing_relative_order_only_creates_branch() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
     env.setup_metadata(&["A"]);
 
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-
-    env.but("move nonexistent A")
+    env.but("status")
         .assert()
-        .failure()
-        .stderr_eq(str![[r#"
-Failed to move commit. Source 'nonexistent' not found. If you just performed a Git operation, try running 'but status' to refresh.
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊●   zll add first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move ywx --above g0")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 9ac4652 to new branch 'a-branch-1' above branch 'A'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ br [a-branch-1]
+┊●   ywx add second
+┊│
+┊├┄ g0 [A]
+┊●   zll add first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
 
 "#]]);
 }
 
 #[test]
-fn move_commit_with_after_flag_and_branch_target() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "commit on A");
-
-    // Get commit CLI ID
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commit_cli_id = status_json["stacks"][0]["branches"][0]["commits"][0]["cliId"]
-        .as_str()
-        .unwrap();
-
-    // Try to move commit to branch with --after flag (should error)
-    env.but(format!("move {commit_cli_id} B --after"))
-        .assert()
-        .failure()
-        .stderr_eq(str![[r#"
-Failed to move commit. The --after flag only makes sense when moving a commit to another commit.
-When moving to a branch, the commit is placed at the top of the stack by default.
-
-"#]]);
-
-    Ok(())
-}
-
-#[test]
-fn move_same_commit_to_itself() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-
+fn moving_commits_below_branch_creates_branch_below() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
     env.setup_metadata(&["A"]);
 
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-
-    // Get commit CLI ID
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commit_cli_id = status_json["stacks"][0]["branches"][0]["commits"][0]["cliId"]
-        .as_str()
-        .unwrap();
-
-    // Move commit to itself (should be no-op)
-    env.but(format!("move {commit_cli_id} {commit_cli_id}"))
+    env.but("status")
         .assert()
         .success()
-        .stdout_eq(str![[r#"
-Source and target are the same commit. Nothing to do.
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊●   zll add first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
 
 "#]]);
 
-    Ok(())
+    env.but("move ywx --below g0")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 9ac4652 to new branch 'a-branch-1' below branch 'A'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   zll add first
+┊│
+┊├┄ br [a-branch-1]
+┊●   ywx add second
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
 }
 
 #[test]
-fn move_commit_with_invalid_target() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-
+fn moving_commits_below_branch_without_changing_relative_order_only_creates_branch() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
     env.setup_metadata(&["A"]);
 
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-
-    // Get commit CLI ID
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commit_cli_id = status_json["stacks"][0]["branches"][0]["commits"][0]["cliId"]
-        .as_str()
-        .unwrap();
-
-    env.but(format!("move {commit_cli_id} nonexistent"))
-        .assert()
-        .failure()
-        .stderr_eq(str![[r#"
-Failed to move commit. Target 'nonexistent' not found. If you just performed a Git operation, try running 'but status' to refresh.
-
-"#]]);
-
-    Ok(())
-}
-
-#[test]
-fn move_cross_stack_works_yay() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
-
-    env.setup_metadata(&["A", "B"]);
-
-    // Create commits on both branches
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "commit on A");
-    commit_two_files_as_two_hunks_each(&env, "B", "c.txt", "d.txt", "commit on B");
-
-    // Get commit CLI IDs
-    let status_output = env.but("--format json status").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-
-    let commit_a_cli_id = status_json["stacks"][0]["branches"][0]["commits"][0]["cliId"]
-        .as_str()
-        .unwrap();
-    let commit_b_cli_id = status_json["stacks"][1]["branches"][0]["commits"][0]["cliId"]
-        .as_str()
-        .unwrap();
-
-    // Try to move commit from A to specific position in B (should work amazingly)
-    env.but(format!("move {commit_a_cli_id} {commit_b_cli_id}"))
+    env.but("status")
         .assert()
         .success()
-        .stdout_eq(str![[r#"
-Moved 1 → before 1
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊●   zll add first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
 
 "#]]);
 
-    Ok(())
+    env.but("move zll --below g0")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved fe12bcd to new branch 'a-branch-1' below branch 'A'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊│
+┊├┄ br [a-branch-1]
+┊●   zll add first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
 }
 
 #[test]
-fn move_committed_file_to_another_commit() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
-
+fn moving_all_commits_above_branch_retains_branch() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
     env.setup_metadata(&["A"]);
 
-    // Create two commits with different files
-    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
-    commit_two_files_as_two_hunks_each(&env, "A", "c.txt", "d.txt", "second commit");
-
-    // Get commit and file CLI IDs from status with files (-f flag)
-    let status_output = env.but("--format json status -f").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commits = &status_json["stacks"][0]["branches"][0]["commits"];
-
-    // commits[0] = "second commit" with c.txt and d.txt
-    // commits[1] = "first commit" with a.txt and b.txt
-    let first_commit_cli_id = commits[1]["cliId"].as_str().unwrap();
-
-    // Get the file CLI ID for c.txt in the second commit
-    // Files are under the "changes" array in the commit
-    let c_txt_file_cli_id = commits[0]["changes"]
-        .as_array()
-        .and_then(|changes| {
-            changes
-                .iter()
-                .find(|change| change["filePath"].as_str() == Some("c.txt"))
-        })
-        .and_then(|change| change["cliId"].as_str())
-        .expect("Could not find c.txt file CLI ID in status output");
-
-    // Move c.txt from second commit to first commit
-    env.but(format!("move {c_txt_file_cli_id} {first_commit_cli_id}"))
+    env.but("status")
         .assert()
         .success()
-        .stdout_eq(str![[r#"
-Moved files between commits!
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊●   zll add first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
 
 "#]]);
 
-    // Verify the file was moved by checking status again
-    let status_output = env.but("--format json status -f").allow_json().output()?;
-    let status_json: serde_json::Value = serde_json::from_slice(&status_output.stdout)?;
-    let commits = &status_json["stacks"][0]["branches"][0]["commits"];
+    env.but("move ywx zll --above g0")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 9ac4652, fe12bcd to new branch 'a-branch-1' above branch 'A'
 
-    // After the move:
-    // - The second commit should now only have d.txt
-    // - The first commit should now have a.txt, b.txt, and c.txt
-    let second_commit_changes = commits[0]["changes"].as_array().unwrap();
-    let first_commit_changes = commits[1]["changes"].as_array().unwrap();
+"#]]);
 
-    assert_eq!(
-        second_commit_changes.len(),
-        1,
-        "Second commit should have 1 file change after moving c.txt"
-    );
-    assert_eq!(
-        first_commit_changes.len(),
-        3,
-        "First commit should have 3 file changes after receiving c.txt"
-    );
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ br [a-branch-1]
+┊●   ywx add second
+┊●   zll add first
+┊│
+┊├┄ g0 [A] (no commits)
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
 
-    Ok(())
+Hint: run `but help` for all commands
+
+"#]]);
 }
 
 #[test]
-fn move_branch_by_name_from_top_level_move() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings(
-        "two-stacks-one-single-and-ready-to-mingle-one-double",
-    );
+fn moving_all_commits_below_branch_retains_branch() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
+    env.setup_metadata(&["A"]);
 
-    env.setup_metadata(&["A", "B"]);
-
-    env.but("st")
+    env.but("status")
         .assert()
         .success()
-        .stderr_eq(snapbox::str![])
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊●   zll add first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move ywx zll --below g0")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 9ac4652, fe12bcd to new branch 'a-branch-1' below branch 'A'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A] (no commits)
+┊│
+┊├┄ br [a-branch-1]
+┊●   ywx add second
+┊●   zll add first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn move_commit_above_empty_branch() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks-one-empty");
+    env.setup_metadata(&["A", "B"]);
+
+    env.but("status")
+        .assert()
+        .success()
         .stdout_eq(snapbox::str![[r#"
 ╭┄ zz [uncommitted] (no changes)
 ┊
@@ -796,11 +574,7 @@ fn move_branch_by_name_from_top_level_move() -> anyhow::Result<()> {
 ┊●   tpm add A
 ├╯
 ┊
-┊╭┄ h0 [C]
-┊●   xwn add C
-┊│
-┊├┄ i0 [B]
-┊●   lrm add B
+┊╭┄ h0 [B] (no commits)
 ├╯
 ┊
 ┴ 0dc3733 (common base) 2000-01-02 add M
@@ -809,89 +583,25 @@ Hint: run `but help` for all commands
 
 "#]]);
 
-    env.but("move A C").assert().success().stdout_eq(str![[r#"
-Moved branch A on top of C.
-
-"#]]);
-
-    let status_json = status_json(&env)?;
-    let layout = stack_branch_layout(&status_json)?;
-    assert_eq!(
-        layout,
-        vec![vec!["A".to_string(), "C".to_string(), "B".to_string()]]
-    );
-
-    env.but("st")
+    env.but("move tpm --above h0")
         .assert()
         .success()
-        .stderr_eq(snapbox::str![])
         .stdout_eq(snapbox::str![[r#"
-╭┄ zz [uncommitted] (no changes)
-┊
-┊╭┄ g0 [A]
-┊●   tpm add A
-┊│
-┊├┄ h0 [C]
-┊●   xwn add C
-┊│
-┊├┄ i0 [B]
-┊●   lrm add B
-├╯
-┊
-┴ 0dc3733 (common base) 2000-01-02 add M
-
-Hint: run `but help` for all commands
+Moved 9477ae7 to new branch 'a-branch-1' above branch 'B'
 
 "#]]);
 
-    Ok(())
-}
-
-#[test]
-fn move_empty_branch_on_top_of_empty_branch_by_cli_id() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
-
-    env.invoke_git("branch A main");
-    env.invoke_git("branch B main");
-    setup_single_stack_metadata(&env, &["B", "A"])?;
-
-    let initial_status = status_json(&env)?;
-    assert_eq!(
-        stack_branch_layout(&initial_status)?,
-        vec![vec!["B".to_string(), "A".to_string()]]
-    );
-
-    let source_branch_cli_id = initial_status["stacks"][0]["branches"][1]["cliId"]
-        .as_str()
-        .context("Missing source branch cliId")?;
-    let target_branch_cli_id = initial_status["stacks"][0]["branches"][0]["cliId"]
-        .as_str()
-        .context("Missing target branch cliId")?;
-
-    env.but(format!(
-        "move {source_branch_cli_id} {target_branch_cli_id}"
-    ))
-    .assert()
-    .success()
-    .stdout_eq(str![[r#"
-Moved branch A on top of B.
-
-"#]]);
-
-    let status_json = status_json(&env)?;
-    assert_eq!(
-        stack_branch_layout(&status_json)?,
-        vec![vec!["A".to_string(), "B".to_string()]]
-    );
-
-    env.but("st")
+    env.but("status")
         .assert()
         .success()
-        .stderr_eq(snapbox::str![])
         .stdout_eq(snapbox::str![[r#"
 ╭┄ zz [uncommitted] (no changes)
 ┊
 ┊╭┄ g0 [A] (no commits)
+├╯
+┊
+┊╭┄ br [a-branch-1]
+┊●   tpm add A
 ┊│
 ┊├┄ h0 [B] (no commits)
 ├╯
@@ -901,49 +611,16 @@ Moved branch A on top of B.
 Hint: run `but help` for all commands
 
 "#]]);
-
-    Ok(())
 }
 
 #[test]
-fn move_branch_onto_itself_fails_without_deleting_branch() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings(
-        "two-stacks-one-single-and-ready-to-mingle-one-double",
-    );
-
+fn move_commit_below_empty_branch() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks-one-empty");
     env.setup_metadata(&["A", "B"]);
 
-    env.but("move C C").assert().failure().stderr_eq(str![[r#"
-Failed to move branch. Cannot move branch refs/heads/C onto itself
-
-"#]]);
-
-    let status_json = status_json(&env)?;
-    let layout = stack_branch_layout(&status_json)?;
-    assert_eq!(
-        layout,
-        vec![
-            vec!["A".to_string()],
-            vec!["C".to_string(), "B".to_string()]
-        ],
-        "self-move failure should leave all branches in their original stacks"
-    );
-
-    Ok(())
-}
-
-#[test]
-fn move_branch_by_cli_id_from_top_level_move() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings(
-        "two-stacks-one-single-and-ready-to-mingle-one-double",
-    );
-
-    env.setup_metadata(&["A", "B"]);
-
-    env.but("st")
+    env.but("status")
         .assert()
         .success()
-        .stderr_eq(snapbox::str![])
         .stdout_eq(snapbox::str![[r#"
 ╭┄ zz [uncommitted] (no changes)
 ┊
@@ -951,11 +628,7 @@ fn move_branch_by_cli_id_from_top_level_move() -> anyhow::Result<()> {
 ┊●   tpm add A
 ├╯
 ┊
-┊╭┄ h0 [C]
-┊●   xwn add C
-┊│
-┊├┄ i0 [B]
-┊●   lrm add B
+┊╭┄ h0 [B] (no commits)
 ├╯
 ┊
 ┴ 0dc3733 (common base) 2000-01-02 add M
@@ -964,46 +637,27 @@ Hint: run `but help` for all commands
 
 "#]]);
 
-    let initial_status = status_json(&env)?;
-    let source_branch_cli_id = initial_status["stacks"][0]["branches"][0]["cliId"]
-        .as_str()
-        .context("Missing source branch cliId")?;
-    let target_branch_cli_id = initial_status["stacks"][1]["branches"][0]["cliId"]
-        .as_str()
-        .context("Missing target branch cliId")?;
-
-    env.but(format!(
-        "move {source_branch_cli_id} {target_branch_cli_id}"
-    ))
-    .assert()
-    .success()
-    .stdout_eq(str![[r#"
-Moved branch A on top of C.
+    env.but("move tpm --below h0")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 9477ae7 to new branch 'a-branch-1' below branch 'B'
 
 "#]]);
 
-    let status_json = status_json(&env)?;
-    let layout = stack_branch_layout(&status_json)?;
-    assert_eq!(
-        layout,
-        vec![vec!["A".to_string(), "C".to_string(), "B".to_string()]]
-    );
-
-    env.but("st")
+    env.but("status")
         .assert()
         .success()
-        .stderr_eq(snapbox::str![])
         .stdout_eq(snapbox::str![[r#"
 ╭┄ zz [uncommitted] (no changes)
 ┊
-┊╭┄ g0 [A]
+┊╭┄ g0 [A] (no commits)
+├╯
+┊
+┊╭┄ h0 [B] (no commits)
+┊│
+┊├┄ br [a-branch-1]
 ┊●   tpm add A
-┊│
-┊├┄ h0 [C]
-┊●   xwn add C
-┊│
-┊├┄ i0 [B]
-┊●   lrm add B
 ├╯
 ┊
 ┴ 0dc3733 (common base) 2000-01-02 add M
@@ -1011,22 +665,16 @@ Moved branch A on top of C.
 Hint: run `but help` for all commands
 
 "#]]);
-
-    Ok(())
 }
 
 #[test]
-fn tear_off_branch_with_top_level_move_to_zz() -> anyhow::Result<()> {
-    let env = Sandbox::init_scenario_with_target_and_default_settings(
-        "two-stacks-one-single-and-ready-to-mingle-one-double",
-    );
-
+fn above_or_below_unapplied_or_non_existing_branch_errors() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks-one-empty");
     env.setup_metadata(&["A", "B"]);
 
-    env.but("st")
+    env.but("status")
         .assert()
         .success()
-        .stderr_eq(snapbox::str![])
         .stdout_eq(snapbox::str![[r#"
 ╭┄ zz [uncommitted] (no changes)
 ┊
@@ -1034,11 +682,7 @@ fn tear_off_branch_with_top_level_move_to_zz() -> anyhow::Result<()> {
 ┊●   tpm add A
 ├╯
 ┊
-┊╭┄ h0 [C]
-┊●   xwn add C
-┊│
-┊├┄ i0 [B]
-┊●   lrm add B
+┊╭┄ h0 [B] (no commits)
 ├╯
 ┊
 ┴ 0dc3733 (common base) 2000-01-02 add M
@@ -1046,35 +690,57 @@ fn tear_off_branch_with_top_level_move_to_zz() -> anyhow::Result<()> {
 Hint: run `but help` for all commands
 
 "#]]);
+    env.but("unapply B").assert().success();
 
-    env.but("move C zz").assert().success().stdout_eq(str![[r#"
-Unstacked branch C.
+    env.but("move tpm --above B")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Could not find anchor: 'B'
+
+Hint: Run `but status` for applicable targets.
 
 "#]]);
 
-    let status_json = status_json(&env)?;
-    let mut all_branches = stack_branch_layout(&status_json)?
-        .into_iter()
-        .flat_map(|stack| {
-            assert_eq!(
-                stack.len(),
-                1,
-                "Each branch should be in its own stack after tear-off"
-            );
-            stack
-        })
-        .collect::<Vec<_>>();
-    all_branches.sort();
+    env.but("move tpm --below B")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Could not find anchor: 'B'
 
-    assert_eq!(
-        all_branches,
-        vec!["A".to_string(), "B".to_string(), "C".to_string()]
-    );
+Hint: Run `but status` for applicable targets.
 
-    env.but("st")
+"#]]);
+
+    env.but("move tpm --above no-such-branch")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Could not find anchor: 'no-such-branch'
+
+Hint: Run `but status` for applicable targets.
+
+"#]]);
+
+    env.but("move tpm --below no-such-branch")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Could not find anchor: 'no-such-branch'
+
+Hint: Run `but status` for applicable targets.
+
+"#]]);
+}
+
+#[test]
+fn move_to_tip_of_branch() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
+    env.setup_metadata(&["A", "B"]);
+
+    env.but("status")
         .assert()
         .success()
-        .stderr_eq(snapbox::str![])
         .stdout_eq(snapbox::str![[r#"
 ╭┄ zz [uncommitted] (no changes)
 ┊
@@ -1086,8 +752,32 @@ Unstacked branch C.
 ┊●   lrm add B
 ├╯
 ┊
-┊╭┄ i0 [C]
-┊●   xwn add C
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move tpm -b B")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 9477ae7 to the tip of branch 'B'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A] (no commits)
+├╯
+┊
+┊╭┄ h0 [B]
+┊●   tpm add A
+┊●   lrm add B
 ├╯
 ┊
 ┴ 0dc3733 (common base) 2000-01-02 add M
@@ -1095,22 +785,16 @@ Unstacked branch C.
 Hint: run `but help` for all commands
 
 "#]]);
-
-    Ok(())
 }
 
 #[test]
-fn move_branch_with_after_flag_fails_from_top_level_move() {
-    let env = Sandbox::init_scenario_with_target_and_default_settings(
-        "two-stacks-one-single-and-ready-to-mingle-one-double",
-    );
-
+fn move_to_tip_of_empty_branch() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks-one-empty");
     env.setup_metadata(&["A", "B"]);
 
-    env.but("st")
+    env.but("status")
         .assert()
         .success()
-        .stderr_eq(snapbox::str![])
         .stdout_eq(snapbox::str![[r#"
 ╭┄ zz [uncommitted] (no changes)
 ┊
@@ -1118,11 +802,7 @@ fn move_branch_with_after_flag_fails_from_top_level_move() {
 ┊●   tpm add A
 ├╯
 ┊
-┊╭┄ h0 [C]
-┊●   xwn add C
-┊│
-┊├┄ i0 [B]
-┊●   lrm add B
+┊╭┄ h0 [B] (no commits)
 ├╯
 ┊
 ┴ 0dc3733 (common base) 2000-01-02 add M
@@ -1131,30 +811,712 @@ Hint: run `but help` for all commands
 
 "#]]);
 
-    env.but("move A C --after")
+    env.but("move tpm -b B")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 9477ae7 to the tip of branch 'B'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A] (no commits)
+├╯
+┊
+┊╭┄ h0 [B]
+┊●   tpm add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn move_to_tip_of_new_unstacked_branch() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
+    env.setup_metadata(&["A"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊●   zll add first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move ywx --branch new-branch")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 9ac4652 to new branch 'new-branch'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ ne [new-branch]
+┊●   ywx add second
+├╯
+┊
+┊╭┄ g0 [A]
+┊●   zll add first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn move_to_tip_of_new_unstacked_branch_with_canned_name() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
+    env.setup_metadata(&["A"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊●   zll add first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move ywx --branch")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 9ac4652 to new branch 'a-branch-1'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ br [a-branch-1]
+┊●   ywx add second
+├╯
+┊
+┊╭┄ g0 [A]
+┊●   zll add first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn move_file_below_commit_creates_commit() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
+    env.setup_metadata(&["A"]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊│     ywx:w A second
+┊●   zll add first
+┊│     zll:l A first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move ywx:wu --below zll")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 1 changes from 9ac4652 to new commit 8e35f84 below commit fe12bcd
+
+"#]]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second (no changes)
+┊●   zll add first
+┊│     zll:l A first
+┊●   1 (no commit message)
+┊│     1:w A second
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn move_file_above_commit_creates_commit() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
+    env.setup_metadata(&["A"]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊│     ywx:w A second
+┊●   zll add first
+┊│     zll:l A first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move zll:lz --above ywx")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 1 changes from fe12bcd to new commit c15e460 above commit 9ac4652
+
+"#]]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   1 (no commit message)
+┊│     1:l A first
+┊●   ywx add second
+┊│     ywx:w A second
+┊●   zll add first (no changes)
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn move_file_below_branch_creates_branch_and_commit() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
+    env.setup_metadata(&["A"]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊│     ywx:w A second
+┊●   zll add first
+┊│     zll:l A first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move ywx:wu --below A")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 1 changes from 9ac4652 to new commit 8e35f84 on new branch 'a-branch-1' below branch 'A'
+
+"#]]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second (no changes)
+┊●   zll add first
+┊│     zll:l A first
+┊│
+┊├┄ br [a-branch-1]
+┊●   1 (no commit message)
+┊│     1:w A second
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn move_file_above_branch_creates_branch_and_commit() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
+    env.setup_metadata(&["A"]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊│     ywx:w A second
+┊●   zll add first
+┊│     zll:l A first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move zll:lz --above A")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 1 changes from fe12bcd to new commit c15e460 on new branch 'a-branch-1' above branch 'A'
+
+"#]]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ br [a-branch-1]
+┊●   1 (no commit message)
+┊│     1:l A first
+┊│
+┊├┄ g0 [A]
+┊●   ywx add second
+┊│     ywx:w A second
+┊●   zll add first (no changes)
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn move_file_to_branch_tip_creates_commit() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
+    env.setup_metadata(&["A", "B"]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   tpm add A
+┊│     tpm:t A A
+├╯
+┊
+┊╭┄ h0 [B]
+┊●   lrm add B
+┊│     lrm:p A B
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move lrm:pl --branch A")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 1 changes from d3e2ba3 to new commit be174de to the tip of branch 'A'
+
+"#]]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   1 (no commit message)
+┊│     1:p A B
+┊●   tpm add A
+┊│     tpm:t A A
+├╯
+┊
+┊╭┄ h0 [B]
+┊●   lrm add B (no changes)
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn move_file_to_non_existing_branch_tip_creates_unstacked_branch_and_commit() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
+    env.setup_metadata(&["A"]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊│     ywx:w A second
+┊●   zll add first
+┊│     zll:l A first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move ywx:wu --branch new-branch")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 1 changes from 9ac4652 to new commit 8e35f84 on new branch 'new-branch'
+
+"#]]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ ne [new-branch]
+┊●   1 (no commit message)
+┊│     1:w A second
+├╯
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second (no changes)
+┊●   zll add first
+┊│     zll:l A first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn move_file_branch_without_argument_creates_unstacked_branch_with_canned_name_and_commit() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
+    env.setup_metadata(&["A"]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊│     ywx:w A second
+┊●   zll add first
+┊│     zll:l A first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move ywx:wu --branch")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 1 changes from 9ac4652 to new commit 8e35f84 on new branch 'a-branch-1'
+
+"#]]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ br [a-branch-1]
+┊●   1 (no commit message)
+┊│     1:w A second
+├╯
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second (no changes)
+┊●   zll add first
+┊│     zll:l A first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn move_file_should_be_order_independent() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
+    env.setup_metadata(&[]);
+
+    env.file("new", "Some data");
+    env.but("commit -m 'Add new file'").assert().success();
+    std::fs::rename(
+        env.projects_root().join("new"),
+        env.projects_root().join("moved"),
+    )
+    .unwrap();
+    env.file("new/file", "Stuff");
+    env.file("unrelated", "This should stay here :)");
+    env.but("commit -m 'Prepare for moves!'").assert().success();
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ br [a-branch-1]
+┊●   1#0 Prepare for moves!
+┊│     1#0:u R moved
+┊│     1#0:p A new/file
+┊│     1#0:t A unrelated
+┊●   1#1 Add new file
+┊│     1#1:n A new
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move 1#0:u 1#0:p --above 1#0")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 2 changes from e3d3e3a to new commit 99ef17e above commit e3d3e3a
+
+"#]]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ br [a-branch-1]
+┊●   1#0 (no commit message)
+┊│     1#0:u R moved
+┊│     1#0:p A new/file
+┊●   1#1 Prepare for moves!
+┊│     1#1:t A unrelated
+┊●   1#2 Add new file
+┊│     1#2:n A new
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("undo").assert().success();
+
+    env.but("move 1#0:p 1#0:u --above 1#0")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 2 changes from e3d3e3a to new commit 99ef17e above commit e3d3e3a
+
+"#]]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ br [a-branch-1]
+┊●   1#0 (no commit message)
+┊│     1#0:u R moved
+┊│     1#0:p A new/file
+┊●   1#1 Prepare for moves!
+┊│     1#1:t A unrelated
+┊●   1#2 Add new file
+┊│     1#2:n A new
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn move_file_from_multiple_source_commits_is_not_allowed() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
+    env.setup_metadata(&["A", "B"]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   tpm add A
+┊│     tpm:t A A
+├╯
+┊
+┊╭┄ h0 [B]
+┊●   lrm add B
+┊│     lrm:p A B
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move tpm:tm lrm:pl -b")
         .assert()
         .failure()
-        .stderr_eq(str![[r#"
-Failed to move branch. The --after flag only makes sense when moving a commit to another commit.
+        .stderr_eq(snapbox::str![[r#"
+Error: Cannot move changes from multiple commits
+
+Hint: Move changes from a single commit at first, then squash additional changes into the new commit
 
 "#]]);
+}
 
-    env.but("st")
+#[test]
+fn move_branch_above_within_same_stack() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings(
+        "one-stack-three-dependent-branches",
+    );
+    env.setup_metadata(&["A", "B", "C"]);
+
+    env.but("status")
         .assert()
         .success()
-        .stderr_eq(snapbox::str![])
         .stdout_eq(snapbox::str![[r#"
 ╭┄ zz [uncommitted] (no changes)
 ┊
-┊╭┄ g0 [A]
+┊╭┄ g0 [C]
+┊●   wlx add C
+┊│
+┊├┄ h0 [B]
+┊●   wwm add B
+┊│
+┊├┄ i0 [A]
 ┊●   tpm add A
 ├╯
 ┊
-┊╭┄ h0 [C]
-┊●   xwn add C
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move B --above C")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Stacked branch 'B' on top of branch 'C'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [B]
+┊●   wwm add B
 ┊│
-┊├┄ i0 [B]
-┊●   lrm add B
+┊├┄ h0 [C]
+┊●   wlx add C
+┊│
+┊├┄ i0 [A]
+┊●   tpm add A
 ├╯
 ┊
 ┴ 0dc3733 (common base) 2000-01-02 add M
@@ -1164,48 +1526,1054 @@ Hint: run `but help` for all commands
 "#]]);
 }
 
-fn status_json(env: &Sandbox) -> anyhow::Result<serde_json::Value> {
-    let output = env.but("--format json status").allow_json().output()?;
-    Ok(serde_json::from_slice(&output.stdout)?)
+#[test]
+#[ignore = "We can't move branches below other branches right now :("]
+fn move_branch_below_within_same_stack() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings(
+        "one-stack-three-dependent-branches",
+    );
+    env.setup_metadata(&["A", "B", "C"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄zz [uncommitted] (no changes)
+┊
+┊╭┄g0 [C]
+┊●   aebb090 add C
+┊│
+┊├┄h0 [B]
+┊●   582f37b add B
+┊│
+┊├┄i0 [A]
+┊●   9477ae7 add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move C --below B")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved branch 'C' below branch 'B'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄zz [uncommitted] (no changes)
+┊
+┊╭┄g0 [B]
+┊●   223f14d add B
+┊│
+┊├┄h0 [C]
+┊●   983f317 add C
+┊│
+┊├┄i0 [A]
+┊●   9477ae7 add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
 }
 
-fn stack_branch_layout(status_json: &serde_json::Value) -> anyhow::Result<Vec<Vec<String>>> {
-    status_json["stacks"]
-        .as_array()
-        .context("Missing stacks array")?
-        .iter()
-        .map(|stack| {
-            stack["branches"]
-                .as_array()
-                .context("Missing branches array")?
-                .iter()
-                .map(|branch| {
-                    branch["name"]
-                        .as_str()
-                        .map(ToOwned::to_owned)
-                        .context("Missing branch name")
-                })
-                .collect::<anyhow::Result<Vec<_>>>()
-        })
-        .collect()
+#[test]
+fn move_branch_above_to_other_stack() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
+    env.setup_metadata(&["A", "B"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   tpm add A
+├╯
+┊
+┊╭┄ h0 [B]
+┊●   lrm add B
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move B --above A")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Stacked branch 'B' on top of branch 'A'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [B]
+┊●   lrm add B
+┊│
+┊├┄ h0 [A]
+┊●   tpm add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
 }
 
-fn setup_single_stack_metadata(env: &Sandbox, branch_names: &[&str]) -> anyhow::Result<()> {
-    let mut meta = env.meta();
-    let mut workspace = meta.workspace(but_core::WORKSPACE_REF_NAME.try_into()?)?;
-    workspace.stacks = vec![WorkspaceStack {
-        id: StackId::from_number_for_testing(0),
-        branches: branch_names
-            .iter()
-            .map(|branch_name| {
-                Ok(WorkspaceStackBranch {
-                    ref_name: format!("refs/heads/{branch_name}").try_into()?,
-                    archived: false,
-                })
-            })
-            .collect::<anyhow::Result<Vec<_>>>()?,
-        workspacecommit_relation: WorkspaceCommitRelation::Merged,
-    }];
-    meta.set_workspace(&workspace)?;
+#[test]
+fn move_empty_branch_above_other_branch() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks-one-empty");
+    env.setup_metadata(&["A", "B"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   tpm add A
+├╯
+┊
+┊╭┄ h0 [B] (no commits)
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move B --above A")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Stacked branch 'B' on top of branch 'A'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [B] (no commits)
+┊│
+┊├┄ h0 [A]
+┊●   tpm add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+#[ignore = "We can't move branches below other branches right now :("]
+fn move_branch_below_to_other_stack() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
+    env.setup_metadata(&["A", "B"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄zz [uncommitted] (no changes)
+┊
+┊╭┄g0 [A]
+┊●   9477ae7 add A
+├╯
+┊
+┊╭┄h0 [B]
+┊●   d3e2ba3 add B
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move A --below B")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved branch 'A' below branch 'B'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄zz [uncommitted] (no changes)
+┊
+┊╭┄g0 [B]
+┊●   e776549 add B
+┊│
+┊├┄h0 [A]
+┊●   9477ae7 add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn unstack_tip_branch() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings(
+        "one-stack-three-dependent-branches",
+    );
+    env.setup_metadata(&["A", "B", "C"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [C]
+┊●   wlx add C
+┊│
+┊├┄ h0 [B]
+┊●   wwm add B
+┊│
+┊├┄ i0 [A]
+┊●   tpm add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move C --unstack")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Unstacked branch 'C'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [B]
+┊●   wwm add B
+┊│
+┊├┄ h0 [A]
+┊●   tpm add A
+├╯
+┊
+┊╭┄ i0 [C]
+┊●   wlx add C
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn unstack_middle_branch() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings(
+        "one-stack-three-dependent-branches",
+    );
+    env.setup_metadata(&["A", "B", "C"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [C]
+┊●   wlx add C
+┊│
+┊├┄ h0 [B]
+┊●   wwm add B
+┊│
+┊├┄ i0 [A]
+┊●   tpm add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move B --unstack")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Unstacked branch 'B'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [B]
+┊●   wwm add B
+├╯
+┊
+┊╭┄ h0 [C]
+┊●   wlx add C
+┊│
+┊├┄ i0 [A]
+┊●   tpm add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn unstack_bottom_branch() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings(
+        "one-stack-three-dependent-branches",
+    );
+    env.setup_metadata(&["A", "B", "C"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [C]
+┊●   wlx add C
+┊│
+┊├┄ h0 [B]
+┊●   wwm add B
+┊│
+┊├┄ i0 [A]
+┊●   tpm add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move A --unstack")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Unstacked branch 'A'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   tpm add A
+├╯
+┊
+┊╭┄ h0 [C]
+┊●   wlx add C
+┊│
+┊├┄ i0 [B]
+┊●   wwm add B
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn unstack_empty_branch() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
+    env.setup_metadata(&[]);
+
+    env.but("branch new bottom").assert().success();
+    env.but("branch new -a bottom top").assert().success();
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ to [top] (no commits)
+┊│
+┊├┄ bo [bottom] (no commits)
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move --unstack top")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Unstacked branch 'top'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ bo [bottom] (no commits)
+├╯
+┊
+┊╭┄ to [top] (no commits)
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn unstack_branch_using_branch_arg() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings(
+        "one-stack-three-dependent-branches",
+    );
+    env.setup_metadata(&["A", "B", "C"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [C]
+┊●   wlx add C
+┊│
+┊├┄ h0 [B]
+┊●   wwm add B
+┊│
+┊├┄ i0 [A]
+┊●   tpm add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    // `--branch` used synonumously with `--unstack`
+    env.but("move A --branch")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Unstacked branch 'A'
+
+"#]]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   tpm add A
+├╯
+┊
+┊╭┄ h0 [C]
+┊●   wlx add C
+┊│
+┊├┄ i0 [B]
+┊●   wwm add B
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+}
+
+#[test]
+fn unstack_file() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
+    env.setup_metadata(&["A"]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊│     ywx:w A second
+┊●   zll add first
+┊│     zll:l A first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move zll --unstack")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved fe12bcd to new branch 'a-branch-1'
+
+"#]]);
+}
+
+#[test]
+fn unstack_commit() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
+    env.setup_metadata(&["A"]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊│     ywx:w A second
+┊●   zll add first
+┊│     zll:l A first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move zll:lz --unstack")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+Moved 1 changes from fe12bcd to new commit d2fbd7a on new branch 'a-branch-1'
+
+"#]]);
+}
+
+#[test]
+fn cannot_unstack_multiple_branches() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings(
+        "one-stack-three-dependent-branches",
+    );
+    env.setup_metadata(&["A", "B", "C"]);
+
+    env.but("move A B --unstack")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Bad input for '<SOURCES>'
+
+Branches can only be moved one at a time
+
+"#]]);
+}
+
+/// This is an API limitation and not a desirable behavior, but moving multiple branches at the same
+/// time is so fringe that it's not worth investing time into right now.
+#[test]
+fn cannot_move_multiple_branches_at_once() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings(
+        "one-stack-three-dependent-branches",
+    );
+    env.setup_metadata(&["A", "B", "C"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [C]
+┊●   wlx add C
+┊│
+┊├┄ h0 [B]
+┊●   wwm add B
+┊│
+┊├┄ i0 [A]
+┊●   tpm add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move A B --above C")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Bad input for '<SOURCES>'
+
+Branches can only be moved one at a time
+
+"#]]);
+}
+
+#[test]
+fn cannot_move_branch_below() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings(
+        "one-stack-three-dependent-branches",
+    );
+    env.setup_metadata(&["A", "B", "C"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [C]
+┊●   wlx add C
+┊│
+┊├┄ h0 [B]
+┊●   wwm add B
+┊│
+┊├┄ i0 [A]
+┊●   tpm add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move C --below B")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Bad input 'B' for '--below'
+
+Invalid target for branch source
+
+Hint: Branches can only be moved with `--above <branch>` to stack or `--unstack` to unstack
+
+"#]]);
+}
+
+#[test]
+fn cannot_move_branch_to_branch_tip() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings(
+        "one-stack-three-dependent-branches",
+    );
+    env.setup_metadata(&["A", "B", "C"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [C]
+┊●   wlx add C
+┊│
+┊├┄ h0 [B]
+┊●   wwm add B
+┊│
+┊├┄ i0 [A]
+┊●   tpm add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move C -b B")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Cannot combine `--branch` with a branch source
+
+"#]]);
+}
+
+#[test]
+fn cannot_mix_sources() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
+    env.setup_metadata(&["A", "B"]);
+
+    env.but("status -f")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   tpm add A
+┊│     tpm:t A A
+├╯
+┊
+┊╭┄ h0 [B]
+┊●   lrm add B
+┊│     lrm:p A B
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move tpm lrm:pl -b")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Bad input for '<SOURCES>'
+
+Mixing source types is not allowed
+
+Hint: You can only move one kind of source (e.g. commits) at a time
+
+"#]]);
+
+    env.but("move lrm B --above A")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Bad input for '<SOURCES>'
+
+Mixing source types is not allowed
+
+Hint: You can only move one kind of source (e.g. commits) at a time
+
+"#]]);
+
+    env.but("move lrm:pl B --above A")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Bad input for '<SOURCES>'
+
+Mixing source types is not allowed
+
+Hint: You can only move one kind of source (e.g. commits) at a time
+
+"#]]);
+}
+
+#[test]
+fn targeting_unapplied_branch_errors() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks-one-empty");
+    env.setup_metadata(&["A", "B"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   tpm add A
+├╯
+┊
+┊╭┄ h0 [B] (no commits)
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+    env.but("unapply B").assert().success();
+
+    env.but("move tpm --branch B")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: A branch named 'B' exists but is not applied
+
+"#]]);
+}
+
+#[test]
+fn cannot_combine_targets() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
+    env.setup_metadata(&["A"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   ywx add second
+┊●   zll add first
+├╯
+┊
+┴ 1bbc04b (common base) 2000-01-02 add Base
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move ywx --below zll --above zll")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+error: the argument '--below <BRANCH_OR_COMMIT>' cannot be used with '--above <BRANCH_OR_COMMIT>'
+
+Usage: but move <--above <BRANCH_OR_COMMIT>|--below <BRANCH_OR_COMMIT>|--branch [<BRANCH>]|--unstack> <SOURCES>...
+
+For more information, try '--help'.
+
+"#]]);
+}
+
+#[test]
+fn must_specify_target() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
+    env.setup_metadata(&[]);
+
+    env.but("move dontcare")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+error: the following required arguments were not provided:
+  <--above <BRANCH_OR_COMMIT>|--below <BRANCH_OR_COMMIT>|--branch [<BRANCH>]|--unstack>
+
+Usage: but move <--above <BRANCH_OR_COMMIT>|--below <BRANCH_OR_COMMIT>|--branch [<BRANCH>]|--unstack> <SOURCES>...
+
+For more information, try '--help'.
+
+"#]]);
+}
+
+#[test]
+fn must_specify_source() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("zero-stacks");
+    env.setup_metadata(&[]);
+
+    env.but("move -b")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+error: the following required arguments were not provided:
+  <SOURCES>...
+
+Usage: but move <--above <BRANCH_OR_COMMIT>|--below <BRANCH_OR_COMMIT>|--branch [<BRANCH>]|--unstack> <SOURCES>...
+
+For more information, try '--help'.
+
+"#]]);
+}
+
+#[test]
+fn source_cannot_be_target() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack-two-commits");
+    env.setup_metadata(&["A"]);
+
+    env.but("move ywx --above ywx")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Bad input 'ywx' for '--above'
+
+Source cannot also be target
+
+Hint: Trying to move items above 'ywx'? Remove 'ywx' from '<SOURCES>' and try again!
+
+"#]]);
+
+    env.but("move ywx --below ywx")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Bad input 'ywx' for '--below'
+
+Source cannot also be target
+
+Hint: Trying to move items below 'ywx'? Remove 'ywx' from '<SOURCES>' and try again!
+
+"#]]);
+
+    env.but("move A --above A")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Bad input 'A' for '--above'
+
+Source cannot also be target
+
+"#]]);
+}
+
+#[test]
+fn cannot_move_from_uncommitted() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
+    env.setup_metadata(&["A"]);
+
+    env.file("file", "some text");
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted]
+┊   qs A file
+┊
+┊╭┄ g0 [A]
+┊●   tpm add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but diff` to see uncommitted changes and `but commit <branch> -m "message" --changes <id>` to commit them
+
+"#]]);
+
+    env.but("move qs -b A")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Bad input 'qs' for '<SOURCES>'
+
+Cannot pass uncommitted file or hunk as source
+
+Hint: A source must be commit, committed file or branch
+
+"#]]);
+    env.but("move zz -b A")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Bad input 'zz' for '<SOURCES>'
+
+Cannot pass uncommitted changes as source
+
+Hint: A source must be commit, committed file or branch
+
+"#]]);
+}
+
+#[test]
+fn cannot_move_to_uncommitted() {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("one-stack");
+    env.setup_metadata(&["A"]);
+
+    env.but("status")
+        .assert()
+        .success()
+        .stdout_eq(snapbox::str![[r#"
+╭┄ zz [uncommitted] (no changes)
+┊
+┊╭┄ g0 [A]
+┊●   tpm add A
+├╯
+┊
+┴ 0dc3733 (common base) 2000-01-02 add M
+
+Hint: run `but help` for all commands
+
+"#]]);
+
+    env.but("move tpm --below zz")
+        .assert()
+        .failure()
+        .stderr_eq(snapbox::str![[r#"
+Error: Expected a commit or a branch, got uncommitted changes
+
+"#]]);
+}
+
+#[test]
+fn move_commit_to_branch_smoke() -> anyhow::Result<()> {
+    let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
+    env.setup_metadata(&["A", "B"]);
+
+    commit_two_files_as_two_hunks_each(&env, "A", "a.txt", "b.txt", "first commit");
+
+    let before = status_json(&env)?;
+    let branch_a_commits_before = branch_commit_cli_ids(&before, "A");
+    let source_cli_id = branch_a_commits_before[0].clone();
+    let branch_b_count_before = branch_commit_cli_ids(&before, "B").len();
+
+    env.but(format!("move {source_cli_id} --branch B"))
+        .assert()
+        .success();
+
+    let after = status_json(&env)?;
+    let branch_a_commits_after = branch_commit_cli_ids(&after, "A");
+    let branch_b_commits_after = branch_commit_cli_ids(&after, "B");
+    assert_eq!(
+        branch_a_commits_after.len() + 1,
+        branch_a_commits_before.len(),
+        "moving one commit should decrease branch A's commit count by one"
+    );
+    assert_eq!(
+        branch_b_commits_after.len(),
+        branch_b_count_before + 1,
+        "moving one commit should increase branch B's commit count by one"
+    );
+    assert!(
+        !branch_a_commits_after.contains(&source_cli_id),
+        "moved commit should no longer be present on branch A"
+    );
+    assert!(
+        branch_b_commits_after.contains(&source_cli_id),
+        "moved commit should be present on branch B"
+    );
+
     Ok(())
 }
