+++ crates/but/tests/but/command/undo/undo_move.rs
@@ -6,7 +6,7 @@ use crate::{
     utils::Sandbox,
 };
 
-// move2: commit to branch
+// move: commit to branch
 #[test]
 fn undo_move_commit_to_branch() {
     let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks");
@@ -21,7 +21,7 @@ fn undo_move_commit_to_branch() {
     let source_cli_id = branch_commit_cli_ids(&status_json(&env).unwrap(), "A")[0].clone();
 
     run_mutate_undo_roundtrip_test(&env, |env| {
-        env.but(format!("_move2 {source_cli_id} --branch B"))
+        env.but(format!("move {source_cli_id} --branch B"))
             .assert()
             .success();
     });
